pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.5",
    "passlib[argon2]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
